pyproject.toml
[project]
name = "data-model-tmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkdocs>=1.6.1,<2",
    "mkdocs-material>=9.7.6",
    "mkdocs-macros-plugin>=1.3.7",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0.3",
]
